changed coral to end effector becaused we removed 2 motors
